Moved save_account to utils module
